Renamed Modal component more accurately and rewrote as a functional component
